Remove the extension prefix for two persistence modules
