+++ gradle/projects.main.properties
@@ -33,8 +33,8 @@ polaris-quarkus-spark-tests=quarkus/spark-tests
 polaris-quarkus-admin=quarkus/admin
 polaris-quarkus-common=quarkus/common
 polaris-quarkus-test-commons=quarkus/test-commons
-polaris-eclipselink=extension/persistence/eclipselink
-polaris-relational-jdbc=extension/persistence/relational-jdbc
+polaris-eclipselink=persistence/eclipselink
+polaris-relational-jdbc=persistence/relational-jdbc
 polaris-tests=integration-tests
 aggregated-license-report=aggregated-license-report
 polaris-immutables=tools/immutables
